feat: add chart label to chartCartesian
