+++ packages/d3fc-chart/examples/chart.js
@@ -30,6 +30,7 @@ var chart = fc.chartCartesian(
 )
   .xLabel('Value')
   .yLabel('Sine / Cosine')
+  .chartLabel('Sine and Cosine')
   .yDomain(yExtent(data))
   .xDomain(xExtent(data))
   .svgPlotArea(multi)
