+++ packages/d3fc-chart/src/cartesianBase.js
@@ -1,5 +1,4 @@
-import { scaleIdentity } from 'd3-scale';
-import { rebindAll, include } from '@d3fc/d3fc-rebind';
+import { rebindAll, include, rebind } from '@d3fc/d3fc-rebind';
 import cartesianChart from './cartesian';
 
 const functor = (v) =>
@@ -8,7 +7,6 @@ const functor = (v) =>
 export default (setPlotArea, defaultPlotArea) =>
     (...args) => {
 
-        let chartLabel = functor('');
         let yLabel = functor('');
         let plotArea = defaultPlotArea;
         let decorate = () => { };
@@ -45,23 +43,6 @@ export default (setPlotArea, defaultPlotArea) =>
                 container.select('.y-label-container>.y-label')
                     .text(yLabel);
 
-                container.select('.top-label')
-                    .style('margin-top', '2em');
-
-                container.enter()
-                    .append('div')
-                    .attr('class', 'chart-label')
-                    .style('grid-column', 3)
-                    .style('-ms-grid-column', 3)
-                    .style('grid-row', 1)
-                    .style('-ms-grid-row', 1)
-                    .style('height', '2em')
-                    .style('line-height', '2em')
-                    .style('text-align', 'center');
-
-                container.select('.chart-label')
-                    .text(chartLabel(data));
-
                 decorate(container, data, index);
             });
 
@@ -69,14 +50,8 @@ export default (setPlotArea, defaultPlotArea) =>
         };
 
         rebindAll(cartesianBase, cartesian, include(/^x/, /^y/));
+        rebind(cartesianBase, cartesian, 'chartLabel');
 
-        cartesianBase.chartLabel = (...args) => {
-            if (!args.length) {
-                return chartLabel;
-            }
-            chartLabel = functor(args[0]);
-            return cartesianBase;
-        };
         cartesianBase.yLabel = (...args) => {
             if (!args.length) {
                 return yLabel;
